+++ test/integration/spec/myaccount/sendRequest.ts
@@ -1,10 +1,35 @@
-import { sendRequest } from '../../../../lib/myaccount/request';
-import { EmailTransaction } from '../../../../lib/myaccount/transactions';
-import { AuthApiError } from '../../../../lib/errors';
-import { createClient, signinAndGetTokens } from '../../util';
 
 // The purpose of this test is to check the "sendRequest" function instead of the backend apis
 // Only test against "/idp/myaccount/emails" endpoint to make sure the function can deliver/throw correct response and error
+
+jest.mock('../../../../lib/features', () => {
+  const actual = jest.requireActual('../../../../lib/features');
+  return {
+    ...actual,
+    isDPoPSupported: () => true
+  };
+});
+
+jest.mock('../../../../lib/oidc/dpop', () => {
+  const actual = jest.requireActual('../../../../lib/oidc/dpop');
+  const keyPair = actual.generateKeyPair();
+
+  return {
+    ...actual,
+    generateKeyPair: async () => await keyPair,
+    findKeyPair: async () => await keyPair,
+    createDPoPKeyPair: async () => {
+      const kp = await keyPair;
+      return  { keyPair: kp, keyPairId: 'foo' };
+    }
+  };
+});
+
+import { sendRequest } from '../../../../lib/myaccount/request';
+import { EmailTransaction } from '../../../../lib/myaccount/transactions';
+import { AuthApiError, AuthSdkError, WWWAuthError } from '../../../../lib/errors';
+import { createClient, signinAndGetTokens } from '../../util';
+
 
 describe('lower level sendRequest function against "/idp/myaccount/emails" endpoint', () => {
   it('can get emails with "okta.myAccount.email.read" token scope', async () => {
@@ -16,8 +41,8 @@ describe('lower level sendRequest function against "/idp/myaccount/emails" endpo
     ];
     const { 
       tokens: { 
-        accessToken: { accessToken } = {}
-      } 
+        accessToken
+      }
     } = await signinAndGetTokens(client, { scopes });
     const transactions = await sendRequest(client, { 
       url: '/idp/myaccount/emails',
@@ -38,6 +63,7 @@ describe('lower level sendRequest function against "/idp/myaccount/emails" endpo
   });
 
   it('throw with 401 when invalid token is provided', async () => {
+    expect.assertions(1);
     const client = createClient({});
     try {
       await sendRequest(client, { 
@@ -46,16 +72,22 @@ describe('lower level sendRequest function against "/idp/myaccount/emails" endpo
         accessToken: 'invalidToken'
       });
     } catch (err) {
-      expect((err as AuthApiError).xhr?.status).toBe(401);
+      if (process.env.USE_DPOP == '1') {
+        expect(err).toBeInstanceOf(AuthSdkError);
+      }
+      else {
+        expect((err as AuthApiError).xhr?.status).toBe(401);
+      }
     }
   });
 
   it('throws 403 when okta.myAccount.email.read no in token scopes', async () => {
+    expect.hasAssertions();
     const client = createClient({});
     const scopes = ['openid', 'profile'];
     const { 
       tokens: { 
-        accessToken: { accessToken } = {}
+        accessToken
       } 
     } = await signinAndGetTokens(client, { scopes });
     try {
@@ -65,7 +97,13 @@ describe('lower level sendRequest function against "/idp/myaccount/emails" endpo
         accessToken 
       });
     } catch (err) {
-      expect((err as AuthApiError).xhr?.status).toBe(403);
+      if (process.env.USE_DPOP == '1') {
+        expect(err).toBeInstanceOf(WWWAuthError);
+        expect((err as WWWAuthError).message).toBe('insufficient_scope');
+      }
+      else {
+        expect((err as AuthApiError).xhr?.status).toBe(403);
+      }
     }
   });
 });